Added predictPopularity
this function uses linear regression to train a model on 4 features of songs in the spotify dataset. I chose these 4 features because they seemed to show the most correlation with popularity.

The function can also return the popularity prediction(0-100) of a random song and compares it to the actual true value for testing purposes.